arreglo error de tipo de datos

+++ Client/src/SdlManager/SdlManager.h
@@ -16,7 +16,7 @@ public:
 private:
     // WIP, deberia ser alguna clase comando
     Queue<int>& commands;                  // alguien me deberia patear esto
-    Queue<std::vector<float>>& positions;  // posiciones para ir actualizando
+    Queue<std::vector<int>>& positions;  // posiciones para ir actualizando
 
     bool event_handler(SdlWorm& worm, SDL_AudioDeviceID device, uint8_t* waveStart,
                        uint32_t waveLength);
